pyproject.toml
[tool.poetry]
name = "github-stats-to-postgres"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
certifi = "^2022.12.7"
charset-normalizer = "^3.1.0"
idna = "^3.4"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
urllib3 = "^1.26.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
